Update oven.py
Changes to the self_heat function to drive an electro-mechanical contactor rather than a SSR.  The original code triggered a switch every duty cycle which is perfect for a zero crossover voltage SSR but not so for a contactor as well as being noisy. The adjusted code is aimed at triggering a switch only when a change in on/off state is called. Seems to work as expected.